docker/requirements.txt
pymongo
python-dotenv
huggingface_hub
transformers
sentence-transformers
torch
deep-translator
pandas
json_repair
sphinx